@@ -14,7 +14,7 @@ topics:
   - native-toolchain
 
 environment:
-  sdk: '>=3.1.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   glob: ^2.1.1
