@@ -6,7 +6,7 @@ import 'dart:io';
 
 import 'package:logging/logging.dart';
 import 'package:native_assets_cli/native_assets_cli.dart'
-    show Architecture, BuildConfig, OS;
+    show Architecture, CCompilerConfig, OS;
 
 import '../native_toolchain/android_ndk.dart';
 import '../native_toolchain/apple_clang.dart';
@@ -21,13 +21,17 @@ import '../tool/tool_instance.dart';
 // TODO(dacoharkes): This should support alternatives.
 // For example use Clang or MSVC on Windows.
 class CompilerResolver {
-  final BuildConfig buildConfig;
+  final CCompilerConfig cCompiler;
+  final OS targetOS;
+  final Architecture? targetArchitecture;
   final Logger? logger;
   final OS hostOS;
   final Architecture hostArchitecture;
 
   CompilerResolver({
-    required this.buildConfig,
+    required this.targetOS,
+    required this.targetArchitecture,
+    required this.cCompiler,
     required this.logger,
     OS? hostOS, // Only visible for testing.
     Architecture? hostArchitecture, // Only visible for testing.
@@ -48,8 +52,6 @@ class CompilerResolver {
       return result;
     }
 
-    final targetOS = buildConfig.targetOS;
-    final targetArchitecture = buildConfig.targetArchitecture;
     final errorMessage =
         "No tools configured on host '${hostOS}_$hostArchitecture' with target "
         "'${targetOS}_$targetArchitecture'.";
@@ -59,17 +61,14 @@ class CompilerResolver {
 
   /// Select the right compiler for cross compiling to the specified target.
   Tool? _selectCompiler() {
-    final targetOS = buildConfig.targetOS;
-    final targetArch = buildConfig.targetArchitecture;
-
     // TODO(dacoharkes): Support falling back on other tools.
-    if (targetArch == hostArchitecture &&
+    if (targetArchitecture == hostArchitecture &&
         targetOS == hostOS &&
         hostOS == OS.linux) return clang;
     if (targetOS == OS.macOS || targetOS == OS.iOS) return appleClang;
     if (targetOS == OS.android) return androidNdkClang;
     if (hostOS == OS.linux) {
-      switch (targetArch) {
+      switch (targetArchitecture) {
         case Architecture.arm:
           return armLinuxGnueabihfGcc;
         case Architecture.arm64:
@@ -84,7 +83,7 @@ class CompilerResolver {
     }
 
     if (hostOS == OS.windows) {
-      switch (targetArch) {
+      switch (targetArchitecture) {
         case Architecture.ia32:
           return clIA32;
         case Architecture.arm64:
@@ -98,15 +97,15 @@ class CompilerResolver {
   }
 
   Future<ToolInstance?> _tryLoadCompilerFromConfig() async {
-    final configCcUri = buildConfig.cCompiler.compiler;
+    final configCcUri = cCompiler.compiler;
     if (configCcUri != null) {
       assert(await File.fromUri(configCcUri).exists());
       logger?.finer('Using compiler ${configCcUri.toFilePath()} '
-          'from BuildConfig.cCompiler.cc.');
+          'from cCompiler.cc.');
       return (await CompilerRecognizer(configCcUri).resolve(logger: logger))
           .first;
     }
-    logger?.finer('No compiler set in BuildConfig.cCompiler.cc.');
+    logger?.finer('No compiler set in cCompiler.cc.');
     return null;
   }
 
@@ -132,8 +131,6 @@ class CompilerResolver {
       return result;
     }
 
-    final targetOS = buildConfig.targetOS;
-    final targetArchitecture = buildConfig.targetArchitecture;
     final errorMessage =
         "No tools configured on host '${hostOS}_$hostArchitecture' with target "
         "'${targetOS}_$targetArchitecture'.";
@@ -143,9 +140,6 @@ class CompilerResolver {
 
   /// Select the right archiver for cross compiling to the specified target.
   Tool? _selectArchiver() {
-    final targetOS = buildConfig.targetOS;
-    final targetArchitecture = buildConfig.targetArchitecture;
-
     // TODO(dacoharkes): Support falling back on other tools.
     if (targetArchitecture == hostArchitecture &&
         targetOS == hostOS &&
@@ -183,15 +177,15 @@ class CompilerResolver {
   }
 
   Future<ToolInstance?> _tryLoadArchiverFromConfig() async {
-    final configArUri = buildConfig.cCompiler.archiver;
+    final configArUri = cCompiler.archiver;
     if (configArUri != null) {
       assert(await File.fromUri(configArUri).exists());
       logger?.finer('Using archiver ${configArUri.toFilePath()} '
-          'from BuildConfig.cCompiler.ar.');
+          'from cCompiler.ar.');
       return (await ArchiverRecognizer(configArUri).resolve(logger: logger))
           .first;
     }
-    logger?.finer('No compiler set in BuildConfig.cCompiler.ar.');
+    logger?.finer('No compiler set in cCompiler.ar.');
     return null;
   }
 
@@ -209,33 +203,31 @@ class CompilerResolver {
       return result;
     }
 
-    final targetOs = buildConfig.targetOS;
-    final targetArchitecture = buildConfig.targetArchitecture;
     final errorMessage =
         "No tools configured on host '${hostOS}_$hostArchitecture' with target "
-        "'${targetOs}_$targetArchitecture'.";
+        "'${targetOS}_$targetArchitecture'.";
     logger?.severe(errorMessage);
     throw ToolError(errorMessage);
   }
 
   Future<ToolInstance?> _tryLoadLinkerFromConfig() async {
-    final configLdUri = buildConfig.cCompiler.linker;
+    final configLdUri = cCompiler.linker;
     if (configLdUri != null) {
       assert(await File.fromUri(configLdUri).exists());
       logger?.finer('Using linker ${configLdUri.toFilePath()} '
-          'from BuildConfig.cCompiler.ld.');
+          'from cCompiler.ld.');
       return (await LinkerRecognizer(configLdUri, hostOS)
               .resolve(logger: logger))
           .first;
     }
-    logger?.finer('No linker set in BuildConfig.cCompiler.ld.');
+    logger?.finer('No linker set in cCompiler.ld.');
     return null;
   }
 
   /// Select the right compiler for cross compiling to the specified target.
   Tool? _selectLinker() {
-    final targetOs = buildConfig.targetOS;
-    final targetArch = buildConfig.targetArchitecture;
+    final targetOs = targetOS;
+    final targetArch = targetArchitecture;
 
     // TODO(dacoharkes): Support falling back on other tools.
     if (targetArch == hostArchitecture &&
@@ -273,7 +265,7 @@ class CompilerResolver {
   }
 
   Future<Uri?> toolchainEnvironmentScript(ToolInstance compiler) async {
-    final fromConfig = buildConfig.cCompiler.envScript;
+    final fromConfig = cCompiler.envScript;
     if (fromConfig != null) {
       logger?.fine('Using envScript from config: $fromConfig');
       return fromConfig;
@@ -287,7 +279,7 @@ class CompilerResolver {
   }
 
   List<String>? toolchainEnvironmentScriptArguments() {
-    final fromConfig = buildConfig.cCompiler.envScriptArgs;
+    final fromConfig = cCompiler.envScriptArgs;
     if (fromConfig != null) {
       logger?.fine('Using envScriptArgs from config: $fromConfig');
       return fromConfig;
