@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:io';
-
 import 'package:logging/logging.dart';
 import 'package:native_assets_cli/native_assets_cli.dart';
 import 'package:native_toolchain_c/native_toolchain_c.dart';
@@ -19,7 +17,10 @@ void main(List<String> arguments) async {
         'src/native_add.c',
         'src/native_multiply.c',
       ],
-      dartBuildFiles: ['hook/build.dart'],
+      dartBuildFiles: [
+        'hook/build.dart',
+        'hook/link.dart',
+      ],
       linkModePreference: LinkModePreference.static,
     );
     await cbuilder.run(
@@ -32,24 +33,27 @@ void main(List<String> arguments) async {
         }),
       linkInPackage: packageName,
     );
-    _placeholderAsset(config, output);
+    output.addAsset(
+      placeholderAsset(config),
+      linkInPackage: packageName,
+    );
   });
 }
 
 // TODO: Currently, linking cannot add assets, so we add a placeholder here.
 // See also https://github.com/dart-lang/native/issues/1084.
-void _placeholderAsset(BuildConfig config, BuildOutput output) {
-  final uri = config.outputDirectory.resolve(
-      config.targetOS.libraryFileName(packageName, DynamicLoadingBundled()));
-  File.fromUri(uri).createSync(recursive: true);
-  output.addAsset(
-    NativeCodeAsset(
-      package: packageName,
-      name: 'src/${packageName}_bindings.dart',
-      linkMode: DynamicLoadingBundled(),
-      file: uri,
-      os: config.targetOS,
-    ),
-    linkInPackage: packageName,
+NativeCodeAsset placeholderAsset(HookConfig config) {
+  final String libraryFileName = config.targetOS.libraryFileName(
+    packageName,
+    DynamicLoadingBundled(),
+  );
+  final uri = config.outputDirectory.resolve(libraryFileName);
+
+  return NativeCodeAsset(
+    package: packageName,
+    name: 'src/${packageName}_bindings.dart',
+    linkMode: DynamicLoadingBundled(),
+    file: uri,
+    os: config.targetOS,
   );
 }
